New methods for BenchmarkRunner to allow for other Mains to use it. Tuning dimple benchmarks.

+++ solvers/java/src/benchmarking/java/com/analog/lyric/dimple/benchmarks/imageDenoising/ImageDenoisingGraph.java
@@ -144,7 +144,7 @@ public class ImageDenoisingGraph
 			{
 				int row = coordinates[0];
 				int col = coordinates[1];
-				return (Double) _vs[row][col].getValue();
+				return (Integer) _vs[row][col].getValue();
 			}
 		}, _rows, _cols);
 	}
